Bit of clean-up
